requirements.txt
requests==2.31.0
google-generativeai==0.8.3
oauthlib==3.2.2
requests-oauthlib==2.0.0
